# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión del inventario de TodoCamisetas. Incluye gestión de Clientes (Regulares/Preferenciales), Tallas y un sistema de cálculo dinámico de precios de Camisetas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Servidor Local Docker"
        }
    ],
    "paths": {
...
        "/sizes": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar todas las tallas",
                "operationId": "getSizes",
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Size"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Crear una nueva talla",
                "operationId": "createSize",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SizeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/sizes/{id}": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Obtener talla por ID",
                "operationId": "getSize",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla encontrada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Actualizar una talla",
                "operationId": "updateSize",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SizeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Eliminar una talla",
                "description": "Falla con 409 si la talla está asociada a una camiseta.",
                "operationId": "deleteSize",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada"
                    },
                    "409": {
                        "description": "Conflicto: talla en uso"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Size": {
                "title": "Size",
                "description": "Modelo de Talla",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "XL"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SizeInput": {
                "title": "Size Input",
                "description": "Datos para crear o actualizar una talla",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "XL"
                    }
                },
                "type": "object"
            }
        }
    },
...
}